turning the classes system to wrappers and containers
starting with the navbar, i will be starting to change the classes system that i have been using for a while (i.e. putting black rect shadow classes on a div gives you a black rectangle with the shadow) into separate classes with their own properties and declarations.

for example: instead of using "white rect shadow button" classes on each navbar button, i could just use the button wrapper on the div the navbar buttons are wrapped in, and it will apply the same style of buttons, as opposed to using the same "white rect shadow button" in all the buttons.

now, the only problem with this is that these are not reusable, BUT it makes working with things WAYYYY easier for me, since i realized that the classes system i created has made it difficult to set different styles to different divs.